--- src/main/java/jpabook/controller/BookForm.java
@@ -1,14 +1,0 @@
-package jpabook.controller;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter @Setter
-public class BookForm {
-    private Long id;
-    private String name;
-    private int price;
-    private int stockQuantity;
-    private String author;
-    private String isbn;
-}
