--- src/main/java/jpabook/controller/ItemController.java
@@ -1,67 +1,0 @@
-package jpabook.controller;
-
-import jpabook.domain.item.Book;
-import jpabook.domain.item.Item;
-import jpabook.service.ItemService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
-import java.util.List;
-@Controller
-@RequiredArgsConstructor
-public class ItemController {
-    private final ItemService itemService;
-
-    @GetMapping(value = "/items/new")
-    public String createForm(Model model) {
-        model.addAttribute("form", new BookForm());
-        return "items/createItemForm";
-    }
-
-    @PostMapping(value = "/items/new")
-    public String create(BookForm form) {
-        Book book = new Book();
-        book.setName(form.getName());
-        book.setPrice(form.getPrice());
-        book.setStockQuantity(form.getStockQuantity());
-        book.setAuthor(form.getAuthor());
-        book.setIsbn(form.getIsbn());
-
-        itemService.saveItem(book);
-        return "redirect:/items";
-    }
-
-    /**
-     * 상품 목록
-     */
-    @GetMapping(value = "/items")
-    public String list(Model model) {
-        List<Item> items = itemService.findItems();
-        model.addAttribute("items", items);
-        return "items/itemList";
-    }
-
-    @GetMapping(value = "/items/{itemId}/edit")
-    public String updateItemForm( @PathVariable("itemId") Long itemId, Model model ) {
-//        Item item = itemService.findItem(itemId);
-        Book item = (Book) itemService.findItem(itemId);
-
-        BookForm form = new BookForm();
-        form.setId( item.getId() );
-        form.setName( item.getName() );
-        form.setPrice( item.getPrice() );
-        form.setStockQuantity( item.getStockQuantity() );
-        form.setAuthor( item.getAuthor() );
-        form.setIsbn( item.getIsbn() );
-
-        model.addAttribute( "form", form );
-        return "items/updateItemForm";
-    }
-
-    @PostMapping(value = "/items/{itemId}/edit")
-    public String updateItem( @ModelAttribute("form") BookForm form, @PathVariable Long itemId ) {
-        itemService.updateItem( itemId, form.getName(), form.getPrice(), form.getStockQuantity() );
-        return "redirect:/items";
-    }
-}
