--- src/main/java/jpabook/controller/MemberController.java
@@ -1,51 +1,0 @@
-package jpabook.controller;
-
-import jakarta.validation.Valid;
-import jpabook.domain.Address;
-import jpabook.domain.Member;
-import jpabook.service.MemberService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-
-import java.util.List;
-
-@Controller
-@RequiredArgsConstructor
-public class MemberController {
-
-    private final MemberService memberService;
-
-    @GetMapping("/members/new")
-    public String createForm(Model model) {
-        model.addAttribute("memberForm", new MemberForm());
-        return "members/createMemberForm";
-    }
-
-    @PostMapping("/members/new")
-    public String create(@Valid MemberForm form, BindingResult result) { // @Valid를 통해 @NotEmpty, @NotNull 등 각 field의 validity를 검증해줌
-
-        if ( result.hasErrors() ) {
-            return "members/createMemberForm";
-        }
-
-        Address address = new Address(form.getCity(), form.getStreet(), form.getZipcode());
-
-        Member member = new Member();
-        member.setName( form.getName() );
-        member.setAddress( address );
-
-        memberService.join( member );
-
-        return "redirect:/"; // 첫 페이지로 redirect
-    }
-
-    @GetMapping("/members")
-    public String list(Model model) {
-        model.addAttribute("members", memberService.findMembers());
-        return "members/memberList";
-    }
-}
