--- src/main/java/jpabook/controller/MemberForm.java
@@ -1,17 +1,0 @@
-package jpabook.controller;
-
-import jakarta.validation.constraints.NotEmpty;
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.stereotype.Controller;
-
-@Getter @Setter
-public class MemberForm {
-
-    @NotEmpty(message = "회원 이름은 필수입니다")
-    private String name;
-
-    private String city;
-    private String street;
-    private String zipcode;
-}
