--- src/main/java/jpabook/domain/OrderStatus.java
@@ -1,5 +1,0 @@
-package jpabook.domain;
-
-public enum OrderStatus {
-    ORDER, CANCEL
-}
