+++ src/main/java/jpabook/jpashop/JpashopApplication.java
@@ -1,4 +1,4 @@
-package jpabook;
+package jpabook.jpashop;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
