+++ src/main/java/jpabook/jpashop/domain/DeliveryStatus.java
@@ -1,4 +1,4 @@
-package jpabook.domain;
+package jpabook.jpashop.domain;
 
 public enum DeliveryStatus {
     READY, COMP
