+++ src/main/java/jpabook/jpashop/domain/OrderItem.java
@@ -1,0 +1,25 @@
+package jpabook.jpashop.domain;
+
+import jakarta.persistence.*;
+import jpabook.jpashop.domain.item.Item;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Getter @Setter
+public class OrderItem {
+
+    @Id @GeneratedValue
+    @Column(name = "order_item_id")
+    private Long id;
+
+    private Item item;
+
+    @ManyToOne
+    @JoinColumn(name = "order_id")
+    private Order order;
+
+    private int orderPrice;
+
+    private int count;
+}
