+++ src/main/java/jpabook/jpashop/domain/OrderStatus.java
@@ -1,0 +1,5 @@
+package jpabook.jpashop.domain;
+
+public enum OrderStatus {
+    READY, COMP
+}
