+++ src/main/java/jpabook/jpashop/domain/item/Album.java
@@ -1,4 +1,4 @@
-package jpabook.domain.item;
+package jpabook.jpashop.domain.item;
 
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
@@ -6,10 +6,10 @@ import lombok.Getter;
 import lombok.Setter;
 
 @Entity
-@DiscriminatorValue("M")
+@DiscriminatorValue("A")
 @Getter @Setter
-public class Movie extends Item {
+public class Album extends Item {
 
-    private String director;
-    private String actor;
+    private String artist;
+    private String etc;
 }
