+++ src/main/java/jpabook/jpashop/domain/item/Movie.java
@@ -1,4 +1,4 @@
-package jpabook.domain.item;
+package jpabook.jpashop.domain.item;
 
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
@@ -6,10 +6,10 @@ import lombok.Getter;
 import lombok.Setter;
 
 @Entity
-@DiscriminatorValue("A")
+@DiscriminatorValue("M")
 @Getter @Setter
-public class Album extends Item {
+public class Movie extends Item {
 
-    private String artist;
-    private String etc;
+    private String director;
+    private String actor;
 }
