--- src/main/java/jpabook/repository/ItemRepository.java
@@ -1,31 +1,0 @@
-package jpabook.repository;
-
-import jakarta.persistence.EntityManager;
-import jpabook.domain.item.Item;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-@RequiredArgsConstructor
-public class ItemRepository {
-
-    private final EntityManager em;
-
-    public void save( Item item ) {
-        if ( item.getId() == null )
-            em.persist( item );
-        else
-            em.merge( item );
-    }
-
-    public Item findOne(Long id) {
-        return em.find(Item.class, id);
-    }
-
-    public List<Item> findAll() {
-        return em.createQuery( "select i from Item i", Item.class )
-                .getResultList();
-    }
-}
