--- src/main/java/jpabook/repository/OrderRepository.java
@@ -1,102 +1,0 @@
-package jpabook.repository;
-
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.TypedQuery;
-import jakarta.persistence.criteria.*;
-import jpabook.domain.Member;
-import jpabook.domain.Order;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-import org.springframework.util.StringUtils;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Repository
-@RequiredArgsConstructor
-public class OrderRepository {
-
-    private final EntityManager em;
-
-    public void save(Order order) {
-        em.persist(order);
-    }
-
-    public Order findOne(Long id) {
-        return em.find( Order.class, id );
-    }
-
-    public List<Order> findAll(OrderSearch orderSearch) {
-
-        // before) 동적쿼리 미적용
-//        return em.createQuery("select o from Order o join o.member m" +
-//                " where o.status = :status" +
-//                " and m.name like :name"
-//                , Order.class)
-//                .setParameter("status", orderSearch.getOrderStatus())
-//                .setParameter("name", orderSearch.getMemberName())
-////                .setFirstResult(10) // 11번째 값부터 시작 (없으면 1번째 값부터(=0))
-//                .setMaxResults(100) // 시작값 기준 100개로 return 제한
-//                .getResultList();\
-
-        // case1) jpql를 문자로 생성
-
-//        //language=JPAQL
-//        String jpql = "select o From Order o join o.member m";
-//        boolean isFirstCondition = true;
-//        //주문 상태 검색
-//        if (orderSearch.getOrderStatus() != null) {
-//            if (isFirstCondition) {
-//                jpql += " where";
-//                isFirstCondition = false;
-//            } else {
-//                jpql += " and";
-//            }
-//            jpql += " o.status = :status";
-//        }
-//        //회원 이름 검색
-//        if (StringUtils.hasText(orderSearch.getMemberName())) {
-//            if (isFirstCondition) {
-//                jpql += " where";
-//                isFirstCondition = false;
-//            } else {
-//                jpql += " and";
-//            }
-//            jpql += " m.name like :name";
-//        }
-//        TypedQuery<Order> query = em.createQuery(jpql, Order.class)
-//                .setMaxResults(1000); //최대 1000건
-//        if (orderSearch.getOrderStatus() != null) {
-//            query = query.setParameter("status", orderSearch.getOrderStatus());
-//        }
-//        if (StringUtils.hasText(orderSearch.getMemberName())) {
-//            query = query.setParameter("name", orderSearch.getMemberName());
-//        }
-//        return query.getResultList();
-
-        // case2) criteria 사용
-        // : jpa criteria를 사용하는 방법으로, JPA에서의 동적query 생성 표준
-
-        CriteriaBuilder cb = em.getCriteriaBuilder();
-        CriteriaQuery<Order> cq = cb.createQuery(Order.class);
-        Root<Order> o = cq.from(Order.class);
-        Join<Order, Member> m = o.join("member", JoinType.INNER); //회원과 조인
-        List<Predicate> criteria = new ArrayList<>();
-        //주문 상태 검색
-        if (orderSearch.getOrderStatus() != null) {
-            Predicate status = cb.equal(o.get("status"),
-                    orderSearch.getOrderStatus());
-            criteria.add(status);
-        }
-        //회원 이름 검색
-        if (StringUtils.hasText(orderSearch.getMemberName())) {
-            Predicate name =
-                    cb.like(m.<String>get("name"), "%" +
-                            orderSearch.getMemberName() + "%");
-            criteria.add(name);
-        }
-        cq.where(cb.and(criteria.toArray(new Predicate[criteria.size()])));
-        TypedQuery<Order> query = em.createQuery(cq).setMaxResults(1000); //최대 1000건
-        return query.getResultList();
-    }
-}
