--- src/main/java/jpabook/repository/OrderSearch.java
@@ -1,12 +1,0 @@
-package jpabook.repository;
-
-import jpabook.domain.OrderStatus;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter @Setter
-public class OrderSearch {
-
-    private String memberName;
-    private OrderStatus orderStatus;
-}
