--- src/main/java/jpabook/service/ItemService.java
@@ -1,39 +1,0 @@
-package jpabook.service;
-
-import jpabook.domain.item.Book;
-import jpabook.domain.item.Item;
-import jpabook.repository.ItemRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-
-@Service
-@Transactional(readOnly = true)
-@RequiredArgsConstructor
-public class ItemService {
-
-    private final ItemRepository itemRepository;
-
-    @Transactional
-    public void saveItem( Item item ) {
-        itemRepository.save(item);
-    }
-
-    @Transactional
-    public void updateItem( Long itemId, String name, int price, int stockQuantity ) {
-        Item findItem = itemRepository.findOne( itemId );
-        findItem.setPrice( price );
-        findItem.setName( name );
-        findItem.setStockQuantity( stockQuantity );
-    }
-
-    public List<Item> findItems() {
-        return itemRepository.findAll();
-    }
-
-    public Item findItem( Long itemId ) {
-        return itemRepository.findOne( itemId );
-    }
-}
