--- src/main/java/jpabook/service/OrderService.java
@@ -1,69 +1,0 @@
-package jpabook.service;
-
-import jpabook.domain.Delivery;
-import jpabook.domain.Member;
-import jpabook.domain.Order;
-import jpabook.domain.OrderItem;
-import jpabook.domain.item.Item;
-import jpabook.repository.ItemRepository;
-import jpabook.repository.MemberRepository;
-import jpabook.repository.OrderRepository;
-import jpabook.repository.OrderSearch;
-import lombok.RequiredArgsConstructor;
-import org.aspectj.weaver.ast.Or;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-
-@Service
-@Transactional(readOnly = true)
-@RequiredArgsConstructor
-public class OrderService {
-
-    private final OrderRepository orderRepository;
-    private final MemberRepository memberRepository;
-    private final ItemRepository itemRepository;
-
-    /**
-     * 주문
-     * @return
-     */
-    @Transactional(readOnly = false)
-    public Long order( Long memberId, Long itemId, int count ) {
-
-        // 엔티티 조회
-        Member member = memberRepository.findOne( memberId );
-        Item item = itemRepository.findOne( itemId );
-
-        // 배송정보 생성
-        Delivery delivery = new Delivery();
-        delivery.setAddress( member.getAddress() );
-
-        // 주문상품 생성
-        OrderItem orderItem = OrderItem.createOrderItem(item, item.getPrice(), count);
-
-        // 주문 생성
-        Order order = Order.createOrder(member, delivery, orderItem);
-
-        // 주문 저장
-        orderRepository.save( order );
-        return order.getId();
-    }
-
-    /**
-     * 취소
-     */
-    @Transactional( readOnly = false )
-    public void cancelOrder(Long orderId) {
-        Order order = orderRepository.findOne(orderId);
-        order.cancel();
-    }
-
-    /**
-     * 검색
-     */
-    public List<Order> findOrders( OrderSearch orderSearch ) {
-        return orderRepository.findAll( orderSearch );
-    }
-}
